Word Cloud
A word cloud is a popular way of visualizing words in text data.
When we applied a word cloud to our dataset, we found that the ‘asia’ and ‘usa’ which were due to the geotagged tweets were frequent. We, therefore, removed them and obtained the following result. 
We could, hence, observe that the words ‘women’, ’men’, ‘harassment’, ’assault’, and ’sexual’ are the most occurring words in the tweets.